requirements.txt
fastapi
uvicorn[standard]
websockets 
python-multipart
supabase
python-dotenv
httpx
pydantic
docker
pytest
psutil
google-generativeai
aiohttp
